`archive`: move rename related function to `rename.go`
